backend/requirements.txt
# FastAPI and web server
fastapi==0.123.5
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.44
alembic==1.17.2
psycopg2-binary==2.9.11

# Task queue
celery==5.6.0
redis==7.1.0

# Authentication
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.3.0

# Configuration & utilities
pydantic-settings==2.12.0
python-dotenv==1.2.1
loguru==0.7.3

# ML libraries for YOLO training
ultralytics==8.3.234
torch==2.9.1
torchvision==0.24.1
opencv-python==4.12.0.88
PyYAML==6.0.3
numpy==2.2.6
scipy==1.16.3
matplotlib==3.10.7
pillow==12.0.0
